fix modal for teams
